SeamCarver.java
The SeamCarver class takes a given picture and modifies it by reducing its size without reducing its content. This class achieves this with a constructor that creates a seam carver object based on the given picture and making a defensive copy of it. It has methods that return the current picture and the width/height of the current picture. It also has methods that can find the shortest horizontal or vertical seam and removing a given horizontal or vertical seam.