Board.java
Board is an immmutable data type that creates a board from a given N-by-N array of tiles and uses this board to implement various
methods. These methods include returning the tile at any position on the board, the suze of the board, the number of tiles out of place, and sum of Manhattan distances between the tiles and the goal board. This API will  also return whehter the input board is the goal board, whether it is solvable, and whether it is equal to another board. Finally, the board iterates throughout all of its neighbor boards, which are all boards obtained by exchanging a tile that is out of place with a neighboring tile.