Autocomplete.java
his API uses both the Term and BinarySearchDeluxe APIs to autocomplete a set of strings and weights. It sorts the terms in lexicographic order, binary searches the query strings that start with a given prefix, and sorts the matching terms in descending order by weight. It creates an Autocomplte data structure and uses two methods to return all terms that start with a given prefix in descending order by weight and to return the number of terms with this given prefix.