Blob.java
This program creates a Blob data type, which keeps track of a Blob's mass, x center of mass and y center of mass and enables pixels to be added into a blob. We can also find the distance between two blobs, which will be used in BeadTracker to track a bead's movement between two consecutive frames.